refactor: remove old contact-preview-panel.jsx
- use new wrapper to compute contact
- use panel to render specific information
